Restore legacy task entrypoints and invoice helpers
